move sequencer_v1_service to core
